优化：自增ID字段升级为bigint-系统字典

+++ sz-common/sz-common-core/src/main/java/com/sz/core/common/entity/DictVO.java
@@ -21,7 +21,7 @@ public class DictVO {
     private String id;
 
     @Schema(description = "字典类型id", example = "1000")
-    private Integer sysDictTypeId;
+    private Long sysDictTypeId;
 
     @Schema(description = "字典名称", example = "正常")
     private String codeName;
